move editor-options types to seperate file
