+++ embeddable-say-editor/shared-types/index.ts
@@ -1,21 +1,3 @@
 /**
  * @file The `shared-types` folder contains types which may be used by both the embeddable-app, and the package files
  */
-
-export type PluginName =
-  | 'citation'
-  | 'article-structure'
-  | 'besluit'
-  | 'besluit-topic'
-  | 'lpdc'
-  | 'roadsign-regulation'
-  | 'variable'
-  | 'table-of-contents'
-  | 'template-comments'
-  | 'confidentiality'
-  | 'location'
-  | 'rdfa-editor'
-  | 'formatting-toggle'
-  | 'html-edit'
-  | 'html-preview'
-  | 'rdfa-blocks-toggle';
